+++ examples/volumetric/streamribbons.py
@@ -15,10 +15,10 @@ pl3d.Update()
 domain = pl3d.GetOutput().GetBlock(0)
 
 ######################## vedo
-box = Mesh(domain, c=None, alpha=0.1)
+box = Mesh(domain, c="white", alpha=0.1)
 
-probe = Line([9,0,28], [11,0,33], res=11).color('k')
+probe = Line([9,0,28], [11,0,33], res=11).color('k').lw(4)
 
-stream = streamLines(domain, probe, direction='backwards', ribbons=2)
+stream = StreamLines(domain, probe, direction='backwards', ribbons=2)
 
 show(box, probe, stream, __doc__, axes=7, bg='bb').close()