+++ docs/changes.md
@@ -23,6 +23,7 @@
 
 ---
 ### `shapes.py`
+- moved `streamLines` to `StreamLines`
 
 ---
 ### `tetmesh.py`
@@ -42,16 +43,25 @@ examples/basic/ssao.py
 examples/basic/align6.py
 examples/basic/pca_ellipsoid.py
 examples/advanced/gyroid.py
+examples/volumetric/volumeOperations.py
+examples/volumetric/streamlines1.py
+
+examples/simulations/orbitals.py
+examples/simulations/mag_field1.py
+examples/simulations/mag_field2.py
+
 examples/pyplot/histo_1d_b.py
 examples/pyplot/fitPolynomial1.py
 examples/pyplot/plot_extra_yaxis.py
-examples/volumetric/volumeOperations.py
+
 examples/other/pygeodesic1.py
+
 
 
 ### Deleted
 examples/other/animation1.py
 examples/other/animation2.py
+
 
 ### Broken
 examples/other/morphomatics_tube.py